requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
wikipedia==1.4.0
pywhatkit
requests==2.31.0
youtube-search-python==1.6.6
langchain==0.1.0
langchain-openai==0.0.2.post1
openai==1.12.0
opencv-python==4.5.5.64
pyautogui==0.9.53
googlesearch-python==1.1.0